LayoutTests/platform/ios-simulator/editing/selection/selection-extend-should-not-move-across-caret-on-mac-expected.txt

On Mac when word-selecting backwards starting with the caret on the middle of a word and then word-selecting forward, the caret is left in the same place where it was, instead of directly selecting to the end of the word (which is windows/unix behavior).

mac:
Extend backward and then forward by word: PASS
Extend forward and then backward by word: PASS
Extend backward and then forward by line: PASS
Extend forward and then backward by line: PASS
Extend backward and then forward by paragraph: PASS
Extend forward and then backward by paragraph: PASS
win:
Extend backward and then forward by word: FAIL: expected "ne%20", got "ne%202%0A"
Extend forward and then backward by word: FAIL: expected "li", got "ne%202%0A"
Extend backward and then forward by line: PASS
Extend forward and then backward by line: PASS
Extend backward and then forward by paragraph: PASS
Extend forward and then backward by paragraph: PASS
unix:
Extend backward and then forward by word: FAIL: expected "ne", got "ne%20"
Extend forward and then backward by word: PASS
Extend backward and then forward by line: PASS
Extend forward and then backward by line: PASS
Extend backward and then forward by paragraph: PASS
Extend forward and then backward by paragraph: PASS
